+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "frontend-scaffold",
+  "name": "c4cneu.com",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
