+++ public/sitemap.xml
@@ -14,6 +14,9 @@
   <url>
     <loc>https://c4cneu.com/apply/brand-designer</loc>
   </url>
+    <url>
+    <loc>https://c4cneu.com/projects</loc>
+  </url>
   <url>
     <loc>https://c4cneu.com/events</loc>
   </url>
