+++ src/pages/apply/developer/ApplyDeveloper.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { makeStyles } from '@material-ui/core/styles';
 import { Container, Box, Typography, Grid } from '@material-ui/core';
-import Hero from '../../components/Hero';
-import { ReactComponent as SVG } from '../../svg/developer.svg';
-import { TextQuoteBlock } from '../../components/TextQuoteBlock';
-import CTA from './CTA';
+import Hero from '../../../components/Hero';
+import { ReactComponent as SVG } from './developer.svg';
+import { TextQuoteBlock } from '../../../components/TextQuoteBlock';
+import CTA from '../CTA';
 // @ts-ignore
 import Fade from 'react-reveal/Fade';
