+++ src/pages/events/Events.tsx
@@ -9,7 +9,7 @@ import {
   Box,
   Link,
 } from '@material-ui/core';
-import { ReactComponent as SVG } from '../../svg/events.svg';
+import { ReactComponent as SVG } from './events.svg';
 import { Link as RouterLink } from 'react-router-dom';
 
 import Hero from '../../components/Hero';
