+++ src/pages/faq/FAQ.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { Container, Box, Link } from '@material-ui/core';
 import Hero from '../../components/Hero';
-import { ReactComponent as SVG } from '../../svg/faq.svg';
+import { ReactComponent as SVG } from './faq.svg';
 import JumpstartAccordion from '../../components/Accordion';
 import { Link as RouterLink } from 'react-router-dom';
 /*
