+++ src/pages/jumpstart/Jumpstart.tsx
@@ -3,7 +3,7 @@ import { Helmet } from 'react-helmet';
 import { makeStyles } from '@material-ui/core/styles';
 import { Container, Box, Grid, Typography, Button } from '@material-ui/core';
 import Hero from '../../components/Hero';
-import { ReactComponent as JumpstartSvg } from '../../svg/jumpstart.svg';
+import { ReactComponent as JumpstartSvg } from './jumpstart.svg';
 import JumpstartAccordion from '../../components/Accordion';
 // @ts-ignore
 import Fade from 'react-reveal/Fade';
