+++ src/pages/people/People.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { makeStyles } from '@material-ui/core/styles';
 import { Container, Grid, Typography } from '@material-ui/core';
-import { ReactComponent as SVG } from '../../svg/people.svg';
+import { ReactComponent as SVG } from './people.svg';
 
 import Member from './Member';
 import Hero from '../../components/Hero';
